refactor: Split API into files
Also use lru_cache instead of app state to get AzureClient and GuacamoleClient
